locmax_issue
